Fix issue with interpolating user code into code template
The problem was that using String#gsub with 2 parameters Ruby will treat
in special way the second parameter. Every '\<digit>' (eg. \1, \2) will
be replaced by appriopriate matching group. We don't want that because
input string is from user and should be replaced literally.

The way to fix it is to use block which returns string. Ruby won't
interpret it in any way.